Remove freenet_ext files not subject to db4o
The ant-contrib and maven dependencies as listed by the removed README
file were needed for libraries unrelated to db4o according to the git
history of the README file. Hence this information is safe to remove.

Removing the FreeBSD JUnit classpath note as well as this is ancient
information and probably outdated.
Furth the db4o build.xml doesn't support JUnit and cannot support it yet
because the necessary db4o unit test library wasn't committed to the
repository.